Clicking on images now enlarges them.  Added download link. Pointed acquistion tab to the reduced images instead of the raw.
