+++ web/sedm_flask.py
@@ -272,8 +272,10 @@ def add_fritz():
     print(request.form)
     if request.data:
         content = json.loads(request.data)
-        # output = open('/scr2/sedm/sedmpy/web/static/fritz_request_%s.txt' % datetime.datetime.utcnow().strftime("%Y%m%d_%H_%M_%S.%f"),'w')
-        output = open('fritz_request_%s.txt' % datetime.datetime.utcnow().strftime(
+
+        #output = open('fritz_request_%s.txt' % datetime.datetime.utcnow().strftime(
+        #    "%Y%m%d_%H_%M_%S.%f"), 'w')
+        output = open('/scr2/sedm/sedmpy/web/static/fritz_request_%s.txt' % datetime.datetime.utcnow().strftime(
             "%Y%m%d_%H_%M_%S.%f"), 'w')
 
         data = json.dumps(content)
@@ -282,14 +284,14 @@ def add_fritz():
         return ('Content-type: text/html\n <title>Accepted Fritz CGI</title>')
     if request.is_json:
         content = json.loads(request.get_json())
-        output = open('fritz_request_%s.txt' % datetime.datetime.utcnow().strftime("%Y%m%d_%H_%M_%S.%f"), 'w')
+        output = open('/scr2/sedm/sedmpy/web/static/fritz_request_%s.txt' % datetime.datetime.utcnow().strftime("%Y%m%d_%H_%M_%S.%f"), 'w')
         data = json.dumps(content)
         output.write(data)
         output.close()
         return ('Content-type: text/html\n <title>Accepted Fritz CGI</title>')
     if request.form:
         content = request.form.to_dict(flat=True)
-        output = open('fritz_request_%s.txt' % datetime.datetime.utcnow().strftime("%Y%m%d_%H_%M_%S.%f"), 'w')
+        output = open('/scr2/sedm/sedmpy/web/static/fritz_request_%s.txt' % datetime.datetime.utcnow().strftime("%Y%m%d_%H_%M_%S.%f"), 'w')
         data = json.dumps(content)
         output.write(data)
         output.close()
@@ -454,7 +456,7 @@ def login_change():
         else:
             return render_template('change_pass.html', sedm_dict=out, form=form)
 
-    return render_template('change_pass.html', sedm_dict={'message': ''}, form=form)
+    return render_template('change_pass.h 000tml', sedm_dict={'message': ''}, form=form)
 
 
 @app.route("/logout")
